[WIP] refactor window
